Update DateTimePicker.vue - change grid layout
Now the grid is filled column by column, stacking elements 1-2-3 in the first column, then the second, and so on. The elements have been re-ordered in the html so that it looks the same as before. But now when tabbing through the elements, up-down arrows for the same number (e.g. year) are sequential, as opposed to tabbing through all the up arrows before getting to all the down arrows.