add blue circle to checkbox focus; also reorder bottom content
Focus styling improved for tab focus accessibility. Now the circle containing the checkboxes highlights with a blue border, similar to the standard color for Chrome.

Also, the date-picker, pause/play time button, and credit logos have been reordered so that the tabbing goes through date-picker, time play, and then credit logos.

Also a couple ordinary changes to the tabbing of some sections of the code for scanability, though the code itself hasn't changed.