separate Nova sheet from use-wwt sheet so they are triggered by two separate buttons
The two elements are called BottomNovaSheet and BottomUseSheet. The original BottomSheet that combined the two and had tabs remains as an element, but is not included in the blaze-star-nova.vue code.
Also, an "i" icon has been added to represent the wwt-usage sheet.
There has also been some adjustments to the tabbing of the code for readability.